requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pymongo>=3.12.1
prettyprint
aiofiles
pycryptodome==3.11.0
motor
requests
# httpx
# pytest-asyncio==0.17.2
# asyncpg==0.22.0

# dev
pytest-asyncio